+++ src/loss.cpp
@@ -1,6 +1,6 @@
 #include "loss.h"
 #include <cmath>
 
-double cross_entropy(std::vector<double> predictions, int target) {
+float cross_entropy(std::vector<float> predictions, int target) {
     return -log(predictions[target]);
 }
