+++ src/neural_network.h
@@ -9,7 +9,7 @@
 
 struct neural_network {
     std::vector<layer> layers;
-    std::function<double(std::vector<double>, std::vector<double>)> error_function;
+    std::function<float(std::vector<float>, std::vector<float>)> error_function;
 public:
     neural_network(std::vector<int> sizes, const std::vector<Activation> &activations);
 
@@ -19,11 +19,11 @@ public:
 
     void forward(const matrix_row_view &input);
 
-    std::vector<double>& get_outputs();
+    std::vector<float>& get_outputs();
 
-    void train(const matrix &inputs, const std::vector<int> &labels, int epochs, double learning_rate);
+    void train(const matrix &inputs, const std::vector<int> &labels, int epochs, float learning_rate);
 
-    void backward(const matrix_row_view &input, const std::vector<double> &label);
+    void backward(const matrix_row_view &input, const std::vector<float> &label);
 
     int accuracy(const matrix &inputs, const std::vector<int> &labels);
 };
